split token request repos into separate files
